Split TestFramework functionality in subclass for failovertest.py
